requirements.txt
flask==2.3.3
openai==1.3.0
psycopg2-binary==2.9.7
pandas==2.1.1
python-dotenv==1.0.0
numpy==1.24.3
scipy==1.11.2
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
gunicorn==21.2.0